Implement Specialty and Cargo Trade Systems

+++ AAEmu.Game/Core/Packets/S2C/TCUccCharNamePacket.cs
@@ -3,7 +3,7 @@ using AAEmu.Game.Core.Network.Stream;
 
 namespace AAEmu.Game.Core.Packets.S2C;
 
-public class TCUccCharNamePacket(uint id, string name) : StreamPacket(TCOffsets.TCUccCharNamePacket)
+public class TCUccCharNamePacket(ulong id, string name) : StreamPacket(TCOffsets.TCUccCharNamePacket)
 {
     public override PacketStream Write(PacketStream stream)
     {
